- upgrade to python3 - test sceleton - refactor - html report integration - insysgo test remove - other improvements
