+++ requirements.txt
@@ -4,7 +4,7 @@ linecache2==1.0.0
 nose==1.3.7
 nose2==0.6.5
 PyYAML==3.12
-selenium==3.4.2
+selenium==3.7.0
 six==1.10.0
 traceback2==1.4.0
 unittest2==1.1.0
