+++ runtests.py
@@ -47,7 +47,7 @@ if __name__ == '__main__':
     if options.result_file:
         result_path = os.path.join("reports", options.result_file)
     else:
-        result_path = os.path.join("reports", "results.xml")
-    test_path = os.path.join("tests", "testcases", "insysgo_tests.py") + os.sep
-    os.system("nosetests-2.7 {test_path} --with-xunit --xunit-file={result_path}"
+        result_path = os.path.join("reports", "results.html")
+    test_path = os.path.join("tests", "testcases", "youtube_tests.py") + os.sep
+    os.system("pytest {test_path} --html={result_path}"
               .format(result_path=result_path, test_path=test_path))
