+++ tests/pages/__init__.py
@@ -2,4 +2,3 @@
 
 # Import your subpages Implementing an Interface in the
 # "__init__.py" so PageFactory will no about it's existence.
-from tests.pages.insysgo_pages import MyListPage
