+++ tests/testdata/settings.py
@@ -1,19 +1,3 @@
-##########################################################################
-# This file is part of WTFramework.
-#
-#    WTFramework is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU General Public License as published by
-#    the Free Software Foundation, either version 3 of the License, or
-#    (at your option) any later version.
-#
-#    WTFramework is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU General Public License for more details.
-#
-#    You should have received a copy of the GNU General Public License
-#    along with WTFramework.  If not, see <http://www.gnu.org/licenses/>.
-##########################################################################
 from wtframework.wtf.config import WTF_CONFIG_READER
 
 # The idea of a 'testdata' package is to organize your test data and test settings.
